backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
cryptography
passlib[bcrypt]
python-multipart
google-auth
requests
redis
celery
kiteconnect
smartapi-python
growwapi
dhanhq
upstox-python-sdk
pyotp
